+++ src/main/java/ch/njol/skript/conditions/base/SingleCondition.java
@@ -1,0 +1,58 @@
+/**
+ *   This file is part of Skript.
+ *
+ *  Skript is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  Skript is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with Skript.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ *
+ * Copyright 2011-2017 Peter Güttinger and contributors
+ */
+package ch.njol.skript.conditions.base;
+
+import org.bukkit.event.Event;
+import org.eclipse.jdt.annotation.Nullable;
+
+import ch.njol.skript.lang.Condition;
+import ch.njol.skript.lang.Expression;
+import ch.njol.skript.lang.SkriptParser;
+import ch.njol.util.Checker;
+import ch.njol.util.Kleenean;
+
+public abstract class SingleCondition<T> extends Condition implements Checker<T> {
+	
+	@SuppressWarnings("null")
+	protected Expression<? extends T> expr;
+	
+	@SuppressWarnings({"unchecked", "null"})
+	@Override
+	public boolean init(final Expression<?>[] exprs, final int matchedPattern, final Kleenean isDelayed, final SkriptParser.ParseResult parseResult) {
+		expr = (Expression<? extends T>) exprs[0];
+		setNegated(matchedPattern == 1);
+		return true;
+	}
+	
+	@Override
+	public final boolean check(final Event e) {
+		return expr.check(e, this, isNegated());
+	}
+	
+	@Override
+	public abstract boolean check(T t);
+	
+	protected abstract String getPropertyName();
+	
+	@Override
+	public String toString(final @Nullable Event e, final boolean debug) {
+		return expr.toString(e, debug) + (expr.isSingle() ? " is " : " are ") + (isNegated() ? "not " : "") + getPropertyName();
+	}
+}
